Backend/requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
Flask-Env==2.0.0
flask-marshmallow==0.9.0
pymysql==0.9.2
mysql-connector==2.2.9
mysql-connector-python==8.0.13
marshmallow-sqlalchemy==0.14.1
click==7.0
pandas==0.25.3
pyserial==3.4
